Removing custom stream applications
Remove custom stream applications provided for partitioning and uppercase transforming.
Going forward, this repository will be used primarily for tesing the out of the box
stream applications on Kubernetes and Cloud Foundry. Therefore, we are removing custom
applications in this repository for better maintenance of testing those out of the box
stream applications.

The partitioning and the custom uppercase applications can still be found under the
partitioning-tests branch.